Some review touchs up
