+++ src/aspire/operators/__init__.py
@@ -14,6 +14,5 @@ from .filters import (
     ScaledFilter,
     ZeroFilter,
     evaluate_src_filters_on_grid,
-    voltage_to_wavelength,
 )
 from .wemd import wemd_embed, wemd_norm
