+++ src/aspire/utils/__init__.py
@@ -65,4 +65,4 @@ from .multiprocessing import (
 from .random import Random, choice, rand, randi, randn, random
 from .rotation import Rotation
 from .types import complex_type, real_type, utest_tolerance
-from .units import ratio_to_decibel
+from .units import ratio_to_decibel, voltage_to_wavelength, wavelength_to_voltage
